refactor(metadata): extract usage-insights queries into launch_queries.rs
Move the 11 query_* methods (inline SQL over launch_operations / launchers /
profiles) from mod.rs into a new launch_queries.rs submodule structured as:

  - 11 pub(super) free functions taking &Connection, each holding the exact
    SQL verbatim
  - one impl MetadataStore block with thin self.with_conn delegations that
    preserve the existing public API (method names, signatures, action strings)

Also extract the in_clause_placeholders helper out of store.rs into a new
util.rs (now called as a free fn from profile_ops.rs and launch_queries.rs
rather than via MetadataStore::in_clause_placeholders).

No behavior change. SQL text moves byte-identical. No schema change.

mod.rs: 2624 → 2238 lines (−386). launch_queries.rs: 457 lines (under the
500-line limit). util.rs: 6 lines. store.rs: 133 → 127 lines.

Phase 3 of 6 for the metadata/mod.rs split.

Refs #291
Refs #290
