+++ src/crosshook-native/crates/crosshook-core/src/metadata/mod.rs
@@ -13,6 +13,7 @@ mod game_image_store;
 mod health_ops;
 mod health_store;
 mod launch_history;
+mod launch_queries;
 mod launcher_ops;
 mod launcher_sync;
 mod migrations;
@@ -33,6 +34,7 @@ mod readiness_dismissal_store;
 mod readiness_snapshot_store;
 mod store;
 mod suggestion_store;
+mod util;
 mod version_ops;
 mod version_store;
 
@@ -53,393 +55,6 @@ pub use proton_catalog_store::ProtonCatalogRow;
 pub use readiness_snapshot_store::HostReadinessSnapshotRow;
 pub use store::MetadataStore;
 pub use version_store::{compute_correlation_status, hash_trainer_file};
-
-use rusqlite::{params_from_iter, OptionalExtension};
-
-impl MetadataStore {
-    // -------------------------------------------------------------------------
-    // Phase 3: Usage insights (inline SQL over launch_operations)
-    // Extracted in Phase 3 of the metadata/mod.rs split refactor. See #291.
-    // -------------------------------------------------------------------------
-
-    pub fn query_most_launched(
-        &self,
-        limit: usize,
-    ) -> Result<Vec<(String, i64)>, MetadataStoreError> {
-        self.with_conn("query most launched profiles", |conn| {
-            let mut stmt = conn
-                .prepare(
-                    "SELECT profile_name, COUNT(*) as launch_count \
-                     FROM launch_operations \
-                     WHERE status IN ('succeeded', 'failed') AND profile_name IS NOT NULL \
-                     GROUP BY profile_name \
-                     ORDER BY launch_count DESC \
-                     LIMIT ?1",
-                )
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_most_launched statement",
-                    source,
-                })?;
-
-            let rows = stmt
-                .query_map(rusqlite::params![limit as i64], |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)?))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_most_launched",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_most_launched row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    pub fn query_total_launches_for_profiles(
-        &self,
-        profile_names: &[String],
-    ) -> Result<Vec<(String, i64)>, MetadataStoreError> {
-        if profile_names.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        self.with_conn("query total launches for profiles", |conn| {
-            let placeholders = Self::in_clause_placeholders(profile_names.len());
-            let sql = format!(
-                "SELECT profile_name, COUNT(*) as launch_count \
-                 FROM launch_operations \
-                 WHERE status IN ('succeeded', 'failed') \
-                   AND profile_name IS NOT NULL \
-                   AND profile_name IN ({placeholders}) \
-                 GROUP BY profile_name"
-            );
-            let mut stmt = conn
-                .prepare(&sql)
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_total_launches_for_profiles statement",
-                    source,
-                })?;
-            let rows = stmt
-                .query_map(params_from_iter(profile_names.iter()), |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)?))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_total_launches_for_profiles",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_total_launches_for_profiles row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    pub fn query_last_success_per_profile(
-        &self,
-    ) -> Result<Vec<(String, String)>, MetadataStoreError> {
-        self.with_conn("query last success per profile", |conn| {
-            let mut stmt = conn
-                .prepare(
-                    "SELECT profile_name, MAX(finished_at) as last_success \
-                     FROM launch_operations \
-                     WHERE status = 'succeeded' AND profile_name IS NOT NULL \
-                     GROUP BY profile_name",
-                )
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_last_success_per_profile statement",
-                    source,
-                })?;
-
-            let rows = stmt
-                .query_map([], |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_last_success_per_profile",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_last_success_per_profile row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    /// Returns `(failures, successes)` for a single profile over the last `days`.
-    /// Missing rows are represented as zero counts.
-    pub fn query_failure_trend_for_profile(
-        &self,
-        profile_name: &str,
-        days: u32,
-    ) -> Result<(i64, i64), MetadataStoreError> {
-        self.with_conn("query failure trend for profile", |conn| {
-            let interval = format!("-{days} days");
-            conn.query_row(
-                "SELECT COUNT(*) FILTER (WHERE status = 'failed') as failures, \
-                        COUNT(*) FILTER (WHERE status = 'succeeded') as successes \
-                 FROM launch_operations \
-                 WHERE started_at >= datetime('now', ?1) AND profile_name = ?2",
-                rusqlite::params![&interval, profile_name],
-                |row| Ok((row.get::<_, i64>(0)?, row.get::<_, i64>(1)?)),
-            )
-            .map_err(|source| MetadataStoreError::Database {
-                action: "query failure trend for profile",
-                source,
-            })
-        })
-    }
-
-    /// Returns the latest succeeded launch timestamp for a single profile.
-    pub fn query_last_success_for_profile(
-        &self,
-        profile_name: &str,
-    ) -> Result<Option<String>, MetadataStoreError> {
-        self.with_conn("query last success for profile", |conn| {
-            conn.query_row(
-                "SELECT MAX(finished_at) \
-                 FROM launch_operations \
-                 WHERE status = 'succeeded' AND profile_name = ?1",
-                rusqlite::params![profile_name],
-                |row| row.get::<_, Option<String>>(0),
-            )
-            .map_err(|source| MetadataStoreError::Database {
-                action: "query last success for profile",
-                source,
-            })
-        })
-    }
-
-    /// Returns the total launches recorded for a single profile.
-    pub fn query_total_launches_for_profile(
-        &self,
-        profile_name: &str,
-    ) -> Result<i64, MetadataStoreError> {
-        self.with_conn("query total launches for profile", |conn| {
-            conn.query_row(
-                "SELECT COUNT(*) \
-                 FROM launch_operations \
-                 WHERE status IN ('succeeded', 'failed') AND profile_name = ?1",
-                rusqlite::params![profile_name],
-                |row| row.get::<_, i64>(0),
-            )
-            .map_err(|source| MetadataStoreError::Database {
-                action: "query total launches for profile",
-                source,
-            })
-        })
-    }
-
-    /// Returns the most recent launcher drift state for a given profile_id,
-    /// or `None` if no launcher rows exist for that profile.
-    pub fn query_launcher_drift_for_profile(
-        &self,
-        profile_id: &str,
-    ) -> Result<Option<DriftState>, MetadataStoreError> {
-        self.with_conn("query launcher drift for profile", |conn| {
-            let result = conn
-                .query_row(
-                    "SELECT drift_state FROM launchers \
-                     WHERE profile_id = ?1 \
-                     ORDER BY updated_at DESC LIMIT 1",
-                    rusqlite::params![profile_id],
-                    |row| row.get::<_, String>(0),
-                )
-                .optional()
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "query launcher drift for profile",
-                    source,
-                })?;
-
-            let drift = result.map(|s| match s.as_str() {
-                "aligned" => DriftState::Aligned,
-                "missing" => DriftState::Missing,
-                "moved" => DriftState::Moved,
-                "stale" => DriftState::Stale,
-                _ => DriftState::Unknown,
-            });
-            Ok(drift)
-        })
-    }
-
-    /// Returns the `source` field from the profiles table for a given profile name,
-    /// or `None` if the profile is not found in the metadata store.
-    pub fn query_profile_source(&self, name: &str) -> Result<Option<String>, MetadataStoreError> {
-        self.with_conn("query profile source", |conn| {
-            conn.query_row(
-                "SELECT source FROM profiles WHERE current_filename = ?1 AND deleted_at IS NULL",
-                rusqlite::params![name],
-                |row| row.get::<_, Option<String>>(0),
-            )
-            .optional()
-            .map(std::option::Option::flatten)
-            .map_err(|source| MetadataStoreError::Database {
-                action: "query profile source",
-                source,
-            })
-        })
-    }
-
-    pub fn query_profile_sources_for_names(
-        &self,
-        profile_names: &[String],
-    ) -> Result<Vec<(String, Option<String>)>, MetadataStoreError> {
-        if profile_names.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        self.with_conn("query profile sources for names", |conn| {
-            let placeholders = Self::in_clause_placeholders(profile_names.len());
-            let sql = format!(
-                "SELECT current_filename, source \
-                 FROM profiles \
-                 WHERE deleted_at IS NULL AND current_filename IN ({placeholders})"
-            );
-            let mut stmt = conn
-                .prepare(&sql)
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_profile_sources_for_names statement",
-                    source,
-                })?;
-            let rows = stmt
-                .query_map(params_from_iter(profile_names.iter()), |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, Option<String>>(1)?))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_profile_sources_for_names",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_profile_sources_for_names row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    pub fn query_launcher_drift_for_profile_ids(
-        &self,
-        profile_ids: &[String],
-    ) -> Result<Vec<(String, DriftState)>, MetadataStoreError> {
-        if profile_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        self.with_conn("query launcher drift for profile ids", |conn| {
-            let placeholders = Self::in_clause_placeholders(profile_ids.len());
-            let sql = format!(
-                "SELECT l.profile_id, l.drift_state \
-                 FROM launchers l \
-                 INNER JOIN ( \
-                   SELECT profile_id, MAX(updated_at) AS max_updated_at \
-                   FROM launchers \
-                   WHERE profile_id IN ({placeholders}) \
-                   GROUP BY profile_id \
-                 ) latest \
-                   ON latest.profile_id = l.profile_id AND latest.max_updated_at = l.updated_at"
-            );
-            let mut stmt = conn
-                .prepare(&sql)
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_launcher_drift_for_profile_ids statement",
-                    source,
-                })?;
-            let rows = stmt
-                .query_map(params_from_iter(profile_ids.iter()), |row| {
-                    let profile_id = row.get::<_, String>(0)?;
-                    let drift_state = row.get::<_, String>(1)?;
-                    let drift = match drift_state.as_str() {
-                        "aligned" => DriftState::Aligned,
-                        "missing" => DriftState::Missing,
-                        "moved" => DriftState::Moved,
-                        "stale" => DriftState::Stale,
-                        _ => DriftState::Unknown,
-                    };
-                    Ok((profile_id, drift))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_launcher_drift_for_profile_ids",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_launcher_drift_for_profile_ids row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    pub fn query_failure_trends(
-        &self,
-        days: u32,
-    ) -> Result<Vec<FailureTrendRow>, MetadataStoreError> {
-        self.with_conn("query failure trends", |conn| {
-            let interval = format!("-{days} days");
-            let mut stmt = conn
-                .prepare(
-                    "SELECT profile_name, \
-                            COUNT(*) FILTER (WHERE status = 'succeeded') as successes, \
-                            COUNT(*) FILTER (WHERE status = 'failed') as failures, \
-                            GROUP_CONCAT(DISTINCT failure_mode) as failure_modes \
-                     FROM launch_operations \
-                     WHERE started_at >= datetime('now', ?1) AND profile_name IS NOT NULL \
-                     GROUP BY profile_name \
-                     HAVING failures > 0 \
-                     ORDER BY failures DESC",
-                )
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_failure_trends statement",
-                    source,
-                })?;
-
-            let rows = stmt
-                .query_map(rusqlite::params![&interval], |row| {
-                    Ok(FailureTrendRow {
-                        profile_name: row.get(0)?,
-                        successes: row.get(1)?,
-                        failures: row.get(2)?,
-                        failure_modes: row.get(3)?,
-                    })
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_failure_trends",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_failure_trends row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-}
 
 #[cfg(test)]
 mod tests {
