+++ src/crosshook-native/crates/crosshook-core/src/metadata/util.rs
@@ -1,0 +1,6 @@
+/// Build a comma-joined list of `?` placeholders for SQL `IN (...)` clauses.
+pub(super) fn in_clause_placeholders(count: usize) -> String {
+    std::iter::repeat_n("?", count)
+        .collect::<Vec<_>>()
+        .join(", ")
+}
